Created checkbox example
