+++ react/components/EXPERIMENTAL_Table/index.tsx
@@ -1,4 +1,5 @@
 import React, { FC, Fragment, forwardRef, PropsWithChildren } from 'react'
+import pick from 'lodash/pick'
 
 import Toolbar from './Toolbar/index'
 import Pagination, { PaginationProps } from './Pagination'
@@ -60,14 +61,33 @@ const Table: RFC<HTMLTableElement, Props> = (
         <Fragment>
           {children}
           <DataTable ref={ref}>
+            <DataTable.Head />
             <DataTable.Body>
               {({ props }) => (
                 <DataTable.Body.Row {...props}>
-                  {({ props, data }) => (
-                    <DataTable.Body.Row.Cell {...props}>
-                      {data}
-                    </DataTable.Body.Row.Cell>
-                  )}
+                  {({ props, data, column, motion }) => {
+                    const { rowHeight, density } = measures
+                    const { id, cellRenderer, condensed, extended } = column
+                    const cellData = condensed
+                      ? pick(data, condensed)
+                      : extended
+                      ? data
+                      : data[id]
+
+                    const content = cellRenderer
+                      ? cellRenderer({
+                          data: cellData,
+                          rowHeight,
+                          density,
+                          motion,
+                        })
+                      : cellData
+                    return (
+                      <DataTable.Body.Row.Cell {...props}>
+                        {content}
+                      </DataTable.Body.Row.Cell>
+                    )
+                  }}
                 </DataTable.Body.Row>
               )}
             </DataTable.Body>
